# openapi-crm-springboot.yaml
openapi: 3.0.3
info:
  title: CRM Backend API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /leads:
    get:
      tags: [Leads]
      summary: List leads
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: assignedUserId
          schema: { type: integer, format: int64 }
        - in: query
          name: fromDate
          schema: { type: string, format: date }
        - in: query
          name: toDate
          schema: { type: string, format: date }
        - in: query
          name: leadType
          schema: { type: string }
        - in: query
          name: leadSource
          schema: { type: string }
      responses:
        '200':
          description: Lead list
    post:
      tags: [Leads]
      summary: Create lead (multipart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LeadCreateMultipart'
      responses:
        '201':
          description: Lead created
  /leads/{id}:
    get:
      tags: [Leads]
      summary: Get lead by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Lead
    put:
      tags: [Leads]
      summary: Update lead (multipart)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LeadUpdateMultipart'
      responses:
        '200':
          description: Lead updated
    delete:
      tags: [Leads]
      summary: Delete lead
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Lead deleted
  /leads/{id}/status:
    patch:
      tags: [Leads]
      summary: Update lead status
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                leadStatus:
                  type: string
      responses:
        '200':
          description: Status updated
  /leads/{id}/convert:
    post:
      tags: [Leads]
      summary: Convert lead to opportunity
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                opportunityName:
                  type: string
      responses:
        '200':
          description: Converted
  /leads/import/indiamart:
    post:
      tags: [Leads]
      summary: Import leads from IndiaMart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fromDate:
                  type: string
                  format: date
                toDate:
                  type: string
                  format: date
      responses:
        '200':
          description: Imported
  /leads/{id}/notes:
    get:
      tags: [Lead Notes]
      summary: List notes by lead id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Notes
    post:
      tags: [Lead Notes]
      summary: Add note to lead
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                leadNote:
                  type: string
      responses:
        '201':
          description: Note added
...
  /leads/{id}/reminders:
    get:
      tags: [Lead Reminders]
      summary: List reminders by lead id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Reminders
    post:
      tags: [Lead Reminders]
      summary: Add reminder to lead
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reminderFor:
                  type: string
                reminderDate:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Reminder added
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    LeadCreateMultipart:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        title: { type: string }
        mobileNo: { type: string }
        address: { type: string }
        email: { type: string }
        city: { type: string }
        state: { type: string }
        country: { type: string }
        phoneNo: { type: string }
        organizationName: { type: string }
        website: { type: string }
        industry: { type: string }
        noOfEmployee: { type: integer }
        source: { type: string }
        type: { type: string }
        reason: { type: string }
        status: { type: string }
        designation: { type: string }
        inquiryDate: { type: string, format: date }
        userIdFk: { type: integer, format: int64 }
        uploadDocument: { type: string, format: binary }
        uploadDocument1: { type: string, format: binary }
        uploadDocument2: { type: string, format: binary }
        uploadDocument3: { type: string, format: binary }
    LeadUpdateMultipart:
      allOf:
        - $ref: '#/components/schemas/LeadCreateMultipart'
        - type: object
          required: []
...